// modules/dxp/apps/osb/osb-site-initializer/osb-site-initializer-partner-portal/extra/remote-app/src/common/enums/mdfClaimColumnKey.ts
/**
 * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
 *
 * The contents of this file are subject to the terms of the Liferay Enterprise
 * Subscription License ("License"). You may not use this file except in
 * compliance with the License. You can obtain a copy of the License by
 * contacting Liferay, Inc. See the License for the specific language governing
 * permissions and limitations under the License, including but not limited to
 * distribution rights of the Software.
 */
export enum MDFClaimColumnKey {
	ACTION = 'FIXED-ACTION',
	AMOUNT_CLAIMED = 'AMOUNT_CLAIMED',
	DATE_SUBMITTED = 'DATE_SUBMITTED',
	MDF_REQUEST_ID = 'MDF_REQUEST_ID',
	PAID = 'PAID',
	PARTNER = 'PARTNER',
	REQUEST_ID = 'REQUEST_ID',
	STATUS = 'STATUS',
	TYPE = 'TYPE',
}