// modules/dxp/apps/osb/osb-site-initializer/osb-site-initializer-partner-portal/extra/remote-app/src/common/enums/mdfColumnKey.ts
/**
 * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
 *
 * The contents of this file are subject to the terms of the Liferay Enterprise
 * Subscription License ("License"). You may not use this file except in
 * compliance with the License. You can obtain a copy of the License by
 * contacting Liferay, Inc. See the License for the specific language governing
 * permissions and limitations under the License, including but not limited to
 * distribution rights of the Software.
 */

export enum MDFColumnKey {
	ACTION = 'FIXED-ACTION',
	TARGET_MARKETS = 'PRMMDFC-006',
	ADDITIONAL_OPTION = 'PRMMDFC-007',
	DATE_SUBMITTTED = 'PRMMDFC-009',
	ID = 'FIXED-ID',
	LAST_MODIFIED = 'PRMMDFC-010',
	OVERALL_CAMPAIGN = 'PRMMDFC-004',
	TARGET_AUDIENCE_ROLES = 'PRMMDFC-008',
	PARTNER = 'PRMMDFC-002',
	LIFERAY_BUSINESS_SALES_GOALS = 'PRMMDFC-005',
	STATUS = 'FIXED-STATUS',
	COUNTRY = 'PRMMDFC-003',
}